Seprate login screen from main
